fixed the cum_sum bug
cum_sum itself is not a problem. The bad access problem occurs when the
sum is NaN. This is because the weight after the exp is way too small.

+++ random_number_gen.c
@@ -60,7 +60,7 @@ randn (double mu, double sigma)
 
 // generate a random number following uniform distribution from [a,b]
 double
-randu(int a, int b)
+randu(double a, double b)
 {
     init();
     //srand((unsigned)time(NULL));
